+++ src/GitHub_Updater/Install.php
@@ -190,14 +190,6 @@ class Install extends Base {
 
 			parent::$options['github_updater_install_repo'] = self::$install['repo'];
 
-			if ( 'plugin' === $type && $this->is_extended_naming() ) {
-				parent::$options['github_updater_install_repo'] = implode( '-', array(
-					self::$install['github_updater_api'],
-					$headers['owner'],
-					self::$install['repo'],
-				) );
-			}
-
 			update_site_option( 'github_updater', Settings::sanitize( parent::$options ) );
 			$url      = self::$install['download_link'];
 			$nonce    = wp_nonce_url( $url );