+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+
+#### 7.1.0 / 2017-08-10
 * always show _Install_ button for single site theme when branch switch is active [#567](https://github.com/afragen/github-updater/issues/567)
 * fixed override of dot org to correctly ignore dot org updates [#581](https://github.com/afragen/github-updater/issues/581)
 * no more extended naming